requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
scapy>=2.5.0
pyftpdlib>=1.5.9
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
